third_party/WebKit/LayoutTests/fast/canvas/webgl/array-set-invalid-arguments-expected.txt

Verifies that attempting to set invalid elements to a Typed Array throws an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS typedArray.set() threw exception TypeError: Cannot read property 'length' of undefined.
PASS typedArray.set('hello world') threw exception RangeError: Source is too large.
PASS typedArray.set(otherArray, 1) threw exception RangeError: Source is too large.
PASS successfullyParsed is true

TEST COMPLETE
